requirements.txt
# AI
datasets
litellm

# General
python-dotenv
